Incorporate feedback #2.

+++ src/plugins/mvcc_delete_plugin.cpp
@@ -1,6 +1,6 @@
 #include "mvcc_delete_plugin.hpp"
 
-#include "numeric"
+#include "concurrency/transaction_manager.hpp"
 #include "operators/get_table.hpp"
 #include "operators/table_wrapper.hpp"
 #include "operators/update.hpp"
@@ -48,7 +48,7 @@ void MvccDeletePlugin::_logical_delete_loop() {
           const CommitID lowest_end_commit_id =
               *std::min_element(std::begin(chunk->mvcc_data()->end_cids), std::end(chunk->mvcc_data()->end_cids));
           const CommitID commit_id_diff = TransactionManager::get().last_commit_id() - lowest_end_commit_id;
-          const CommitID max_commit_id_diff =
+          auto max_commit_id_diff =
               static_cast<CommitID>(table->max_chunk_size() * _DELETE_THRESHOLD_COMMIT_DIFF_FACTOR);
           const bool criterion2 = max_commit_id_diff <= commit_id_diff;
 
@@ -78,23 +78,26 @@ void MvccDeletePlugin::_logical_delete_loop() {
 void MvccDeletePlugin::_physical_delete_loop() {
   std::unique_lock<std::mutex> lock(_mutex_physical_delete_queue);
 
-  TableAndChunkID table_and_chunk_id = _physical_delete_queue.front();
-  const auto& table = table_and_chunk_id.first;
-  const auto& chunk = table->get_chunk(table_and_chunk_id.second);
+  if (_physical_delete_queue.size()) {
+    TableAndChunkID table_and_chunk_id = _physical_delete_queue.front();
+    const auto& table = table_and_chunk_id.first;
+    const auto& chunk = table->get_chunk(table_and_chunk_id.second);
 
-  DebugAssert(chunk != nullptr, "Chunk does not exist. Physical Delete can not be applied.");
+    DebugAssert(chunk != nullptr, "Chunk does not exist. Physical Delete can not be applied.");
 
-  if (chunk->get_cleanup_commit_id().has_value()) {
-    // Check whether there are still active transactions that might use the chunk
-    bool conflicting_transactions = false;
-    auto lowest_snapshot_commit_id = TransactionManager::get().get_lowest_active_snapshot_commit_id();
-    if (lowest_snapshot_commit_id.has_value()) {
-      conflicting_transactions = chunk->get_cleanup_commit_id().value() < lowest_snapshot_commit_id.value();
-    }
+    if (chunk->get_cleanup_commit_id().has_value()) {
+      // Check whether there are still active transactions that might use the chunk
+      bool conflicting_transactions = false;
+      auto lowest_snapshot_commit_id = TransactionManager::get().get_lowest_active_snapshot_commit_id();
 
-    if (!conflicting_transactions) {
-      _delete_chunk_physically(table, table_and_chunk_id.second);
-      _physical_delete_queue.pop();
+      if (lowest_snapshot_commit_id.has_value()) {
+        conflicting_transactions = chunk->get_cleanup_commit_id().value() > lowest_snapshot_commit_id.value();
+      }
+
+      if (!conflicting_transactions) {
+        _delete_chunk_physically(table, table_and_chunk_id.second);
+        _physical_delete_queue.pop();
+      }
     }
   }
 }
@@ -112,9 +115,10 @@ bool MvccDeletePlugin::_try_logical_delete(const std::string& table_name, const 
   auto gt = std::make_shared<GetTable>(table_name);
   gt->set_transaction_context(transaction_context);
 
-  std::vector<ChunkID> excluded_chunk_ids(table->chunk_count());
-  std::iota(std::begin(excluded_chunk_ids), std::end(excluded_chunk_ids), 0);
-  excluded_chunk_ids.erase(std::remove(std::begin(excluded_chunk_ids), std::end(excluded_chunk_ids), chunk_id));
+  // Include all ChunksIDs of current table except chunk_id for pruning in GetTable
+  std::vector<ChunkID> excluded_chunk_ids(table->chunk_count() - 1);
+  std::iota(excluded_chunk_ids.begin(), excluded_chunk_ids.begin() + chunk_id, 0);
+  std::iota(excluded_chunk_ids.begin() + chunk_id, excluded_chunk_ids.end(), chunk_id + 1);
 
   gt->set_excluded_chunk_ids(excluded_chunk_ids);
   gt->execute();
@@ -147,9 +151,9 @@ bool MvccDeletePlugin::_try_logical_delete(const std::string& table_name, const 
 void MvccDeletePlugin::_delete_chunk_physically(const std::shared_ptr<Table>& table, const ChunkID chunk_id) {
   const auto& chunk = table->get_chunk(chunk_id);
 
-  Assert(chunk.use_count() == 2,
-         "At this point, the chunk should be referenced by the plugin and the "
-         "Table-chunk-vector only.");
+  Assert(chunk.use_count() == 3,
+         "At this point, the chunk should be referenced only by the plugin in methods _delete_chunk_physically() "
+         "& _physical_delete_loop() and the table's chunk vector.");
   Assert(chunk->get_cleanup_commit_id().has_value(),
          "The cleanup commit id of the chunk is not set. "
          "This should have been done by the logical delete.");
